requirements.txt
install==1.3.5
psycopg2-binary==2.9.9
# psycopg2==2.9.9
python-dotenv==1.0.1
fastapi 
uvicorn
